Rename joblib/test into joblib/tests
for consistency with generally used convention, e.g. numpy, scipy,
scikit-learn.
