--- joblib/test/__init__.py
@@ -1,2 +1,0 @@
-from joblib.test import test_memory
-from joblib.test import test_hashing
