+++ joblib/tests/__init__.py
@@ -1,0 +1,2 @@
+from joblib.tests import test_memory
+from joblib.tests import test_hashing
