+++ joblib/tests/test_format_stack.py
