+++ joblib/tests/test_my_exceptions.py
